third_party/WebKit/LayoutTests/inspector/elements/styles/style-formatter-expected.txt

Tests the script formatting functionality.


Running: testScriptFormatterWorker

Running: testSourceMapping

Running: testFormatInlinedStyles
<html><body><style>
@-webkit-keyframes {
**from {
****left: 0
**}

**to {
****left: 100px;
**}
}
</style><style>
badbraces {
}
}

@media screen {
**a {
****color: red;
****text-decoration: none
**}
}
</style></body></html>
